requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
sentence-transformers
chromadb
groq
pypdf